Move ParsedTopHits to test code
This is only used in InternalAggregationTestCase and one of its
children, no need for it to live in prod code.
